sistema_vuelos/requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
psycopg2-binary==2.9.9
bcrypt==4.0.1
gunicorn==21.2.0
